pyproject.toml
[tool.poetry]
name = "itmo-proj-fall-pt"
version = "0.1.0"
description = ""
authors = ["Новый ФизТех <edu@physics.itmo.ru>"]
readme = "README.md"
license = "MIT"
packages = [
    {include = "src"},
    {include = "TRIM"}
]

include = [
    "LICENSE",
]

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
PyQt6 = "^6.3.1"
numpy = "^1.23.3"
pyqtgraph = "^0.13.1"
numpy-stl = "^2.17.1"
pyopengl = "^3.1.6"
pandas = "^1.5.2"
openpyxl = "^3.0.10"
rsinstrument = "^1.53.0"
pywin32 = "^305"
matplotlib = "^3.7.1"
SQLAlchemy = "^2.0.9"
SQLAlchemy-Utils = "^0.41.0"
mat4py = "^0.5.0"
scipy = "^1.10.1"
pyinstaller = "^5.11.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.3"
pyinstaller = "^5.6.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
